Add 1.2 multiplier for weather boost
scripts.js checks if weather boost is enabled and if the current type is boosted and multiplies by 1.2. Note that matchups are still not sorted by multiplier, and matchups other than 0.625 and 1.6 are in strong text.

Weather boost toggle and weather boost select update the type display (now necessary)

weatherboost.js now includes a getWeather method that returns the currently selected weather