Move to svelte kit
Finally got around to completely moving this repo to svelte  / svelte kit with typescript.

Currently, everything has been moved, but the website is broken due to missing links/components that need to be moved to svelte components. Since this change is massive enough as is, committing this as one change. Since I'm leaving the /docs directory, the existing site should work fine even if I push this. Will eventually be able to remove docs but only after the move to Netlify is complete. Will take some precision and attention to detail to avoid downtime.